Added 'exclude_jquery' optional argument.
Because Plack::Middleware::Debug includes an older version of jQuery and
loads into the page by replacing the closing `</body>` tag, any existing
jQuery object is replaced.  This is somewhat mitigated by using
`jQuery.noConflict();` in `toolbar.js`, but still results in
'window.jQuery' referencing the older `1.2.6` version of jQuery that is
loaded from `debug_toolbar/jquery.js`.

If jQuery plugins were added prior to `debug_toolbar/jquery.js`, then
they are no longer available under `window.jQuery` afterwards, as
`noConflict()` simply restores the reference `$` previously pointed to,
and the variable `jQuery` conintues to point to the `1.2.6` version.

This can also cause problems for code executed via `$(document).ready()`
functions that reference `jQuery` (rather than `$`) when calling modern
methods that are not available in the older `1.2.6` version of jQuery.

A longer term solution would be to rewrite `toolbar.js` to remove its
dependence on jQuery, but this small change to add 'exclude_jquery'
makes it easy to work around the problem in the meantime.